+++ admin/common.js
@@ -1,0 +1,11 @@
+const packageJson = require('../package.json');
+
+
+/**
+ * Get the version.
+ */
+const getVersion = () => packageJson && packageJson.version;
+
+module.exports = {
+  getVersion,
+};
