+++ utils/lang/lang.js
@@ -1,9 +1,7 @@
 /**
  * Validates if a value is an object.
  */
-export function isObject(obj) {
-  return obj && typeof obj === 'object' && obj.constructor === Object;
-}
+const isObject = (obj) => obj && typeof obj === 'object' && obj.constructor === Object;
 
 module.exports = {
   isObject,
